% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_freq_values}
\alias{get_freq_values}
\title{Transform frequency text into separate rows}
\usage{
get_freq_values(sample_freq)
}
\arguments{
\item{sample_freq}{A string containing frequency data in the format:
"width,min_length,freq1,freq2,...(size,freq)..."}
}
\value{
A data frame with two columns:
\itemize{
\item size: numeric, representing the size class
\item frequency: numeric, representing the frequency for each size class
}
}
\description{
This function takes a string containing frequency data in a specific format and
transforms it into a data frame with size and frequency information.
}
\examples{
get_freq_values("0.5,7.5,1,2,1")
get_freq_values("0.5,7.5,1,+10,2")
}
