% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harvesting.R
\name{harvest_excel_data_stat}
\alias{harvest_excel_data_stat}
\title{Harvest Excel data from statistical catch and effort sheets and save as CSV files}
\usage{
harvest_excel_data_stat(input_directory, file_name, output_directory)
}
\arguments{
\item{input_directory}{Directory containing the input Excel file}

\item{file_name}{Name of the Excel file to process}

\item{output_directory}{Directory where CSV files will be saved}
}
\value{
None. The function saves processed data as CSV files in the output directory.
}
\description{
Harvest Excel data from statistical catch and effort sheets and save as CSV files
}
\details{
This function processes an Excel file containing statistical catch and effort data.
It assumes the Excel file has two worksheets: 'catch' and 'effort'.
The function reads these worksheets, processes the data, and saves the results as CSV files.
}
