% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{read_sheet}
\alias{read_sheet}
\title{Read a sheet from an Excel file}
\usage{
read_sheet(excel_file, sheet_name, stat_data = FALSE)
}
\arguments{
\item{excel_file}{Path to the Excel file}

\item{sheet_name}{Name of the sheet to read}

\item{stat_data}{If TRUE, all columns are read as text. Default is FALSE.}
}
\value{
A data frame containing the data from the specified Excel sheet
}
\description{
Read a sheet from an Excel file
}
